+++ source/isaaclab/isaaclab/envs/mdp/commands/__init__.py
@@ -8,12 +8,13 @@
 from .commands_cfg import (
     NormalVelocityCommandCfg,
     NullCommandCfg,
+    ObstaclePose2dCommandCfg,
     TerrainBasedPose2dCommandCfg,
     UniformPose2dCommandCfg,
     UniformPoseCommandCfg,
     UniformVelocityCommandCfg,
 )
 from .null_command import NullCommand
-from .pose_2d_command import TerrainBasedPose2dCommand, UniformPose2dCommand
+from .pose_2d_command import ObstaclePose2dCommand, TerrainBasedPose2dCommand, UniformPose2dCommand
 from .pose_command import UniformPoseCommand
 from .velocity_command import NormalVelocityCommand, UniformVelocityCommand