+++ source/isaaclab_rl/isaaclab_rl/rsl_rl/__init__.py
@@ -16,5 +16,5 @@ The following example shows how to wrap an environment for RSL-RL:
 """
 
 from .exporter import export_policy_as_jit, export_policy_as_onnx
-from .rl_cfg import RslRlOnPolicyRunnerCfg, RslRlPpoActorCriticCfg, RslRlPpoAlgorithmCfg
+from .rl_cfg import RslRlOnPolicyRunnerCfg, RslRlPpoActorCriticCfg, RslRlPpoActorCriticConv2dCfg, RslRlPpoAlgorithmCfg
 from .vecenv_wrapper import RslRlVecEnvWrapper